Update delete function in UserController
Update delete function in UserController to also delete all user notifications, events
and remove user from following, followers and requests on all users.
Also remove the users session and remove set user of metaUser to undefined

+++ src/Controllers/UserController.js
@@ -1,9 +1,11 @@
 import uuidv4 from 'uuid/v4';
 import * as User from '../lib/User';
 import { hashPassword } from '../Helpers/db';
-import { createMetaUser } from '../lib/MetaUser';
-import { createSession } from '../lib/Session';
+import { createMetaUser, handleMetaUserUpdate } from '../lib/MetaUser';
+import { createSession, deleteSession } from '../lib/Session';
 import { uploadProfileImageToCloudinary, invalidImage } from '../Helpers/cloudinary';
+import { deleteEvents } from '../lib/Event';
+import { deleteNotifications } from '../lib/Notification';
 
 export default {
   async create(req, res) {
@@ -118,10 +120,74 @@ export default {
   async delete(req, res) {
     if (req.params.userId && !req.isAdmin) return res.status(403).json({ error: 'Forbidden', message: 'You are not Authorzied to delete this user' });
     const userId = req.params.userId ? req.params.userId : req.userId;
+    try {
+      const user = await User.findUserById(userId).catch(error => error);
 
-    const user = await User.deleteUser({ _id: userId }).catch(error => error);
+      if (user.errmsg) return res.status(500).json({ error: user, message: 'Error occurred when trying to delete the user' });
 
-    if (user.errmsg) return res.status(500).json({ error: user, message: 'Error occurred when trying to delete the user' });
+      if (user.notifications.length > 0) {
+        const query = { _id: { $in: user.notifications } };
+        const notifications = await deleteNotifications(query).catch(error => error);
+
+        if (notifications.errmsg) return res.status(500).json({ error: notifications, message: 'Error occurred when trying to delete the user' });
+      }
+      if (user.events.length > 0) {
+        const query = { _id: { $in: user.events } };
+        const events = await deleteEvents(query).catch(error => error);
+        if (events.errmsg) return res.status(500).json({ error: events, message: 'Error occurred when trying to delete the user' });
+      }
+
+      if (user.followers.length > 0) {
+        await Promise.all(user.followers.map(async (follower) => {
+          const updatedUser = await User.handleUserUpdate(follower, { following: userId });
+
+          if (updatedUser.errmsg) return res.status(500).json({ error: updatedUser, message: 'Error occurred when trying to delete the user' });
+
+          return follower;
+        }));
+      }
+
+      if (user.following.length > 0) {
+        await Promise.all(user.following.map(async (following) => {
+          const updatedUser = await User.handleUserUpdate(following, { followers: userId });
+
+          if (updatedUser.errmsg) return res.status(500).json({ error: updatedUser, message: 'Error occurred when trying to delete the user' });
+
+          return following;
+        }));
+      }
+
+      const query = { requests: userId };
+
+      const userWithRequests = await User.findUsers({ query }).catch(error => error);
+
+      if (userWithRequests.length > 0) {
+        await Promise.all(userWithRequests.map(async (requestedUser) => {
+          const updatedUser = await User.handleUserUpdate(requestedUser._id, { requests: userId });
+
+          if (updatedUser.errmsg) return res.status(500).json({ error: updatedUser, message: 'Error occurred when trying to delete the user' });
+
+          return requestedUser;
+        }));
+      }
+
+      const session = await deleteSession({ user: userId }).catch(error => error);
+
+      if (session.errmsg) return res.status(500).json({ error: session, message: 'Error occurred when trying to delete the user' });
+
+      const updatedMetaUser = (
+        await handleMetaUserUpdate(user.metaUser, { user: undefined })
+          .catch(error => error)
+      );
+
+      if (updatedMetaUser.errmsg) return res.status(500).json({ error: updatedMetaUser, message: 'Error occurred when trying to delete the user' });
+
+      const deletedUser = await User.deleteUser({ _id: userId }).catch(error => error);
+
+      if (deletedUser.errmsg) return res.status(500).json({ error: deletedUser, message: 'Error occurred when trying to delete the user' });
+    } catch (e) {
+      return res.status(500).json({ error: e, message: 'Error occurred when trying to delete the user' });
+    }
 
     return res.status(200).json({ info: 'User was deleted' });
   },
